Google sheets update, data collection and analysis
Fixed plant-analysis so it works with both EtherCalc and Google Sheets. The data collection from Phidgets now has two versions, one for EtherCalc and one for Google Sheets. Removed the automatic posting of data in order to improve performance. The students can collect data with a button push. There is optional code to enable automatic data collection every 15 minutes, if the student or teacher is willing to experiment. (Works on local Jupyter Hubs and on 2i2c. Fails on hub.callysto.ca)