Replace versions_0 with full versions, side note
The basic xxx_0 versions were too similar to the full version. So I just got rid of them to make the folder simpler. Mary agrees this is a good idea, so I went ahead with it. The side note in each file explains how to use the anywidget directly, without adding any Python code or links to ipywidgets.